requirements.txt
python-dotenv
pydantic
google-generativeai
reportlab
streamlit 
google-genai 
requests
